requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# LLM Providers (Claude via Anthropic is primary)
anthropic==0.18.1
openai==1.12.0  # Optional fallback

# Web Search
tavily-python==0.3.3

# Async Support
httpx==0.26.0
aiohttp==3.9.3

# Utilities
tenacity==8.2.3

